+++ ui/admin/src/views/instance-misp.vue
@@ -122,7 +122,7 @@ export default {
         },
         delete_instance(elem)
         {
-            axios.get(`/api/misp/delete/${elem.id}`, { timeout: 10000, headers: {'X-Token': this.jwt} })
+            axios.delete(`/api/misp/delete/${elem.id}`, { timeout: 10000, headers: {'X-Token': this.jwt} })
             .then(response => {
                 if(response.data.status){
                     this.instances = this.instances.filter(function(el) { return el != elem; });