+++ ui/admin/src/views/whitelist-search.vue
@@ -67,9 +67,9 @@ export default {
             return true;
         },
         remove: function(elem){
-            axios.get(`/api/whitelist/delete/${elem.id}`, { 
-                timeout: 10000, 
-                headers: {'X-Token': this.jwt} 
+            axios.delete(`/api/whitelist/delete/${elem.id}`, {
+                timeout: 10000,
+                headers: {'X-Token': this.jwt}
             }).then(response => {
                 if(response.data.status){
                     this.results = this.results.filter(function(el) { return el != elem; });